requirements.txt
django-watson
BeautifulSoup==3.2.1
Django==1.6.2
PyYAML==3.10
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
bleach==1.4
django-admin-tools==0.5.1
django-allauth==0.16.1
django-ipware==0.0.8
html5lib==0.999

nose==1.3.1
oauthlib==0.6.1
pygeoip==0.3.1
pyparsing==2.0.1
python-dateutil==2.2
python-openid==2.2.5
requests==2.2.1
requests-oauthlib==0.4.0
six==1.5.2
tornado==3.2
wsgiref==0.1.2
